Move as optional authenticated-proxy feature
