+++ scripts/patch-linux-window-ui.js
@@ -84,7 +84,6 @@ const {
   applyLinuxLocalAppServerFeatureEnablementHandlerPatch,
   applyLinuxOpaqueBackgroundPatch,
   applyLinuxOwlFeatureBindingFallbackPatch,
-  applyLinuxProxyAuthPatch,
   applyLinuxQuitGuardPatch,
   applyLinuxReadyToShowWindowStatePatch,
   applyLinuxResizeRepaintPatch,
@@ -241,7 +240,6 @@ module.exports = {
   applyLinuxMultiInstanceBootstrapPatch,
   applyLinuxOpaqueBackgroundPatch,
   applyLinuxOwlFeatureBindingFallbackPatch,
-  applyLinuxProxyAuthPatch,
   applyLinuxOpaqueWindowsDefaultPatch,
   applyLinuxFastModeModelGuardPatch,
   applyLinuxQuitGuardPatch,
