+++ scripts/patch-linux-window-ui.test.js
@@ -60,7 +60,6 @@ const {
   applyLinuxAppServerBackfillWaitPatch,
   applyLinuxOpaqueBackgroundPatch,
   applyLinuxOwlFeatureBindingFallbackPatch,
-  applyLinuxProxyAuthPatch,
   applyLinuxFastModeModelGuardPatch,
   applyLinuxOpaqueWindowsDefaultPatch,
   applyLinuxReadyToShowWindowStatePatch,
@@ -699,7 +698,6 @@ test("default core patch descriptors are grouped and unique", () => {
     "linux-tray",
     "linux-build-info-tray",
     "linux-single-instance",
-    "linux-proxy-auth",
     "linux-computer-use-ui-feature",
     "linux-computer-use-plugin-gate",
     "linux-chrome-plugin-auto-install",
@@ -2818,190 +2816,6 @@ test("adds Linux single-instance lock and second-instance handoff", () => {
   assert.match(patched, /codexLinuxSecondInstanceHandler/);
   assert.match(patched, /n\.app\.on\(`second-instance`,codexLinuxSecondInstanceHandler\)/);
   assert.match(patched, /n\.app\.off\(`second-instance`,codexLinuxSecondInstanceHandler\)/);
-});
-
-test("registers Linux proxy authentication before Electron app ready", async () => {
-  const source = [
-    "let n=require(`electron`);",
-    "async function boot(){",
-    "t.Er().info(`Launching app`,{safe:{agentRunId:process.env.CODEX_ELECTRON_AGENT_RUN_ID?.trim()||null}});",
-    "let A=Date.now();",
-    "await n.app.whenReady();",
-    "return A}",
-  ].join("");
-  const patched = applyPatchTwice(applyLinuxProxyAuthPatch, source);
-
-  assert.match(patched, /function codexLinuxInstallProxyAuthHandler\(e\)/);
-  assert.match(patched, /codexLinuxInstallProxyAuthHandler\(n\);await n\.app\.whenReady\(\)/);
-
-  const handlers = {};
-  const calls = { whenReady: 0 };
-  const app = {
-    on(event, handler) {
-      handlers[event] = handler;
-    },
-    whenReady() {
-      calls.whenReady += 1;
-      return Promise.resolve();
-    },
-  };
-  const context = {
-    Date,
-    process: {
-      platform: "linux",
-      env: {
-        CODEX_LINUX_PROXY_AUTH_HOST: "proxy.example",
-        CODEX_LINUX_PROXY_AUTH_PORT: "8080",
-        CODEX_LINUX_PROXY_USERNAME: "user",
-        CODEX_LINUX_PROXY_PASSWORD: "p@ss",
-      },
-    },
-    require(name) {
-      assert.equal(name, "electron");
-      return { app };
-    },
-    t: {
-      Er() {
-        return { info() {} };
-      },
-    },
-  };
-
-  await vm.runInNewContext(`${patched};boot()`, context);
-  assert.equal(calls.whenReady, 1);
-  assert.equal(typeof handlers.login, "function");
-
-  let prevented = 0;
-  let credentials = null;
-  handlers.login(
-    { preventDefault() { prevented += 1; } },
-    null,
-    null,
-    { isProxy: true, host: "PROXY.EXAMPLE", port: 8080 },
-    (username, password) => {
-      credentials = { username, password };
-    },
-  );
-
-  assert.equal(prevented, 1);
-  assert.deepEqual(credentials, { username: "user", password: "p@ss" });
-
-  credentials = null;
-  handlers.login(
-    { preventDefault() { prevented += 1; } },
-    null,
-    null,
-    { isProxy: false, host: "proxy.example", port: 8080 },
-    (username, password) => {
-      credentials = { username, password };
-    },
-  );
-
-  assert.equal(prevented, 1);
-  assert.equal(credentials, null);
-});
-
-test("routes authenticated proxy desktop fetches through ClientRequest login", async () => {
-  const source = [
-    "let a=require(`electron`);",
-    "async function boot(){await a.app.whenReady()}",
-    "class Fetcher{",
-    "async performDesktopFetch({body:e,headers:n,method:r,onUploadProgress:i,resolvedUrl:o,signal:s}){let p=()=>e,m=async e=>{let t=this.cloneHeaders(n);let d=i==null?await a.net.fetch(o,{method:r,headers:t,body:p(),signal:s}):await this.performProgressRequest({body:p(),headers:t,method:r,onUploadProgress:i,resolvedUrl:o,signal:s});return d};return m({})}",
-    "performProgressRequest({body:e,headers:t,method:n,onUploadProgress:r,resolvedUrl:i,signal:o}){return new Promise((s,c)=>{let l=a.net.request({method:n,url:i,headers:t}),u=-1,d=()=>{let e=l.getUploadProgress();!e.started||e.current===u||(u=e.current,r({loaded:e.current,total:e.total}))};if(o.addEventListener(`abort`,()=>{},{}),o.aborted)return;l.on(`error`,e=>c(e)),l.on(`response`,e=>{d();let t=[];e.on(`data`,e=>{t.push(e)}),e.on(`end`,()=>{let n=Buffer.concat(t),r=new Headers;for(let[t,n]of Object.entries(e.headers))for(let e of Array.isArray(n)?n:[n])r.append(t,e);s(new Response(n.length===0?null:n,{status:e.statusCode,statusText:e.statusMessage,headers:r}))})});let g=e instanceof ArrayBuffer?Buffer.from(e):e;l.end(g)})}",
-    "cloneHeaders(e){return e}",
-    "}",
-    "globalThis.Fetcher=Fetcher;",
-  ].join("");
-  const patched = applyPatchTwice(applyLinuxProxyAuthPatch, source);
-
-  assert.match(patched, /function codexLinuxAttachProxyAuthToRequest\(e\)/);
-  assert.match(patched, /i==null&&!codexLinuxProxyAuthEntry\(\)\?await a\.net\.fetch/);
-  assert.match(
-    patched,
-    /codexLinuxAttachProxyAuthToRequest\(l\);let u=-1,d=\(\)=>\{if\(r==null\)return;/,
-  );
-
-  let fetchCalls = 0;
-  let requestCalls = 0;
-  let credentials = null;
-  const app = {
-    on() {},
-    whenReady() {
-      return Promise.resolve();
-    },
-  };
-  const net = {
-    fetch() {
-      fetchCalls += 1;
-      return Promise.resolve(new Response("fetch"));
-    },
-    request() {
-      requestCalls += 1;
-      const handlers = {};
-      return {
-        on(event, handler) {
-          handlers[event] = handler;
-          return this;
-        },
-        getUploadProgress() {
-          return { started: false, current: 0, total: 0 };
-        },
-        end() {
-          handlers.login?.(
-            { isProxy: true, host: "PROXY.EXAMPLE", port: 8080 },
-            (username, password) => {
-              credentials = { username, password };
-            },
-          );
-          const responseHandlers = {};
-          handlers.response?.({
-            statusCode: 200,
-            statusMessage: "OK",
-            headers: { "content-type": "text/plain" },
-            on(event, handler) {
-              responseHandlers[event] = handler;
-              return this;
-            },
-          });
-          responseHandlers.data?.(Buffer.from("request"));
-          responseHandlers.end?.();
-        },
-      };
-    },
-  };
-  const context = {
-    ArrayBuffer,
-    Buffer,
-    DOMException,
-    Headers,
-    Response,
-    globalThis: {},
-    process: {
-      platform: "linux",
-      env: {
-        CODEX_LINUX_PROXY_AUTH_HOST: "proxy.example",
-        CODEX_LINUX_PROXY_AUTH_PORT: "8080",
-        CODEX_LINUX_PROXY_USERNAME: "user",
-        CODEX_LINUX_PROXY_PASSWORD: "p@ss",
-      },
-    },
-    require(name) {
-      assert.equal(name, "electron");
-      return { app, net };
-    },
-  };
-
-  await vm.runInNewContext(`${patched};boot()`, context);
-  const response = await vm.runInNewContext(
-    "new globalThis.Fetcher().performDesktopFetch({body:null,headers:{},method:`GET`,resolvedUrl:`https://chatgpt.com/wham/usage`,signal:{aborted:false,addEventListener(){},removeEventListener(){}}})",
-    context,
-  );
-
-  assert.equal(fetchCalls, 0);
-  assert.equal(requestCalls, 1);
-  assert.equal(response.status, 200);
-  assert.equal(await response.text(), "request");
-  assert.deepEqual(credentials, { username: "user", password: "p@ss" });
 });
 
 test("forces the bootstrap single-instance lock on Linux even when upstream disables it", () => {
