--- scripts/patches/core/all-linux/main-process/network-proxy/patch.js
@@ -1,15 +1,0 @@
-"use strict";
-
-const {
-  applyLinuxProxyAuthPatch,
-} = require("../../../../main-process.js");
-
-module.exports = [
-  {
-    id: "linux-proxy-auth",
-    phase: "main-bundle",
-    order: 125,
-    ciPolicy: "optional",
-    apply: applyLinuxProxyAuthPatch,
-  },
-];
